fix(tooling): keep Trellis overlay guard self-contained
Move the runner-guard implementation into
.omp/extensions/enso-subagent-guard so OMP/Bun no longer loads src/.
The overlay factory imports only ./guard and host packages, and throws
if neither ExtensionRunner specifier resolves.
